Removed obsolete API endpoint and merged 2 pages.
