+++ FribergFastigheter.Client/Pages/BrokerDetailsPage.razor
@@ -1,3 +1,15 @@
-﻿@page "/broker/{id:int}"
-
-<h3>Broker</h3>
+﻿@page "/broker/{BrokerId:int}"
+@using FribergFastigheter.Client.Components
+@*
+<!-- Author: Jimmie -->
+<!-- Co Authors: -->
+*@
+<div class="vertical-page-container">
+	@if (Broker != null && BrokerHousings != null)
+	{
+		<div class="d-flex flex-column gap-5">
+			<BrokerDetails Broker="Broker"></BrokerDetails>
+			<HousingList Housings="BrokerHousings"></HousingList>
+		</div>
+	}
+</div>
