--- FribergFastigheter.Client/Pages/HousingBrokerPage.razor
@@ -1,15 +1,0 @@
-﻿@page "/Housing/{housingId:int}/Broker"
-@using FribergFastigheter.Client.Components
-@* 
-<!-- Author: Jimmie -->
-<!-- Co Authors: --> 
-*@
-<div class="vertical-page-container">
-	@if (Broker != null && BrokerHousings != null)
-	{
-		<div class="d-flex flex-column gap-5">
-			<BrokerDetails Broker="Broker"></BrokerDetails>
-			<HousingList Housings="BrokerHousings"></HousingList>
-		</div>
-	}	
-</div>
